# refactor: make `WithEscapeHatch<T>` much more performant