--- a/packages/generator/src/artifacts/types/prop-types.ts
+++ b/packages/generator/src/artifacts/types/prop-types.ts
@@ -24,12 +24,12 @@ export function generatePropTypes(ctx: Context) {
   return outdent`
   ${result.join('\n')}
 
-  type WithColorOpacityModifier<T> = T extends string ? \`$\{T}/\${string}\` : T
+  type WithColorOpacityModifier<T> = [T] extends [string] ? \`$\{T}/\${string}\` & { __colorOpacityModifier?: true } : never
 
   type ImportantMark = "!" | "!important"
   type WhitespaceImportant = \` \${ImportantMark}\`
   type Important = ImportantMark | WhitespaceImportant
-  type WithImportant<T> = T extends string ? \`\${T}\${Important}\` & { __important?: true } : T;
+  type WithImportant<T> = [T] extends [string] ? \`\${T}\${Important}\` & { __important?: true } : never
 
   /**
    * Only relevant when using \`strictTokens\` or \`strictPropertyValues\` in your config.