This pull request is a follow-up pull request to #2466.

## 📝 Description

Make `WithEscapeHatch<T>` much more performant and typescript happy by updating the type signature of `WithImportant<T>` and `WithColorOpacityModifier<T>` to use *branded type* and *non-distributive conditional types*, while keeping such tokens valid and also not appearing in autocompletions to prevent them from polluting autocompletion result (which is the current behavior).

## ⛳️ Current behavior (updates)

Current type signaure of `WithImportant<T>` and `WithColorOpacityModifier<T>` somehow makes typescript compiler unhappy and slows down both code autocompletion and linting.

## 🚀 New behavior

New type signaure of `WithImportant<T>` and `WithColorOpacityModifier<T>` is much more performant. On using `svelte-check` in large codebase was taking over 3 minutes but now it takes under 4 seconds. VSCode autocompletion is also nearly instant, instead of taking ~5 seconds to show.

## 💣 Is this a breaking change (Yes/No):

No.

## 📝 Additional Information