@devunt 2024-08-25
## Benchmark code
Runned by `@arktype/attest@0.9.4` with `bun@1.1.26`, `typescript@5.3.3`.

```ts
import { bench } from '@arktype/attest'

import { css } from '../styled-system-strict/css'

bench('bench type', () => {
  css({
    display: 'flex',
    flexDirection: 'column!',
    alignItems: 'center !important',
    color: 'amber.100/10',
    backgroundColor: 'red.100/50!',
    borderColor: 'blue.100/90 !important',
  })
})
  .mean()
  .types()
```


## Original (baseline):
```
🏌️  bench type
Call: bench type (mean):
⛳ Result: 630.31ns
🎯 Baseline: 619.55ns
📊 Delta: +1.74%

undefined
⛳ Result: 10702 instantiations
🎯 Baseline: 10702 instantiations
📊 Delta: 0.00%

---

🏌️  bench type
Call: bench type (mean):
⛳ Result: 631.27ns
🎯 Baseline: 619.55ns
📊 Delta: +1.89%

undefined
⛳ Result: 10702 instantiations
🎯 Baseline: 10702 instantiations
📊 Delta: 0.00%

---

🏌️  bench type
Call: bench type (mean):
⛳ Result: 638.70ns
🎯 Baseline: 619.55ns
📊 Delta: +3.09%

undefined
⛳ Result: 10702 instantiations
🎯 Baseline: 10702 instantiations
📊 Delta: 0.00%
```

## After modification (This PR):
```
🏌️  bench type
Call: bench type (mean):
⛳ Result: 620.07ns
🎯 Baseline: 619.55ns
📊 Delta: +0.08%

undefined
⛳ Result: 4287 instantiations
🎯 Baseline: 10702 instantiations
📉 bench type was under baseline by 59.94%! Consider setting a new baseline.

---

took 6s at 18:10:08
🏌️  bench type
Call: bench type (mean):
⛳ Result: 627.87ns
🎯 Baseline: 619.55ns
📊 Delta: +1.34%

undefined
⛳ Result: 4287 instantiations
🎯 Baseline: 10702 instantiations
📉 bench type was under baseline by 59.94%! Consider setting a new baseline.

---

🏌️  bench type
Call: bench type (mean):
⛳ Result: 630.60ns
🎯 Baseline: 619.55ns
📊 Delta: +1.78%

undefined
⛳ Result: 4287 instantiations
🎯 Baseline: 10702 instantiations
📉 bench type was under baseline by 59.94%! Consider setting a new baseline.
```

## Conclusion
We can see significant drop in type instaniations count which will reduce overall type check performance drastically.